pyproject.toml
[tool.poetry]
name = "megaqc"
version = "0.3.0"
description = "Collect and visualise data across multiple MultiQC runs"
authors = ["Phil Ewels"]
license = "GPLv3"
readme = "README.rst"
keywords = [
        "bioinformatics",
        "biology",
        "sequencing",
        "NGS",
        "next generation sequencing",
        "quality control",
]
homepage = "https://megaqc.info/"
repository = "https://github.com/MultiQC/MegaQC"
documentation = "https://megaqc.info/docs/index.html"
classifiers = [
        "Development Status :: 4 - Beta",
        "Environment :: Console",
        "Environment :: Web Environment",
        "Intended Audience :: Science/Research",
        "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
        "Natural Language :: English",
        "Operating System :: MacOS :: MacOS X",
        "Operating System :: POSIX",
        "Operating System :: Unix",
        "Programming Language :: Python",
        "Programming Language :: JavaScript",
        "Topic :: Scientific/Engineering",
        "Topic :: Scientific/Engineering :: Bio-Informatics",
        "Topic :: Scientific/Engineering :: Visualization",
]
packages = [{ include = "megaqc" }]

[tool.poetry.dependencies]
python = "^3.9"
argon2-cffi = ">=20.1"
click = "^8.0"
# FastAPI and related
fastapi = ">=0.100.0"
uvicorn = { version = ">=0.23.0", extras = ["standard"] }
python-multipart = ">=0.0.6"
starlette = ">=0.27.0"
# Jinja2 for templates
Jinja2 = "^3.0"
# Database
SQLAlchemy = "^2.0"
alembic = "^1.12"
# Authentication
python-jose = { version = "^3.3", extras = ["cryptography"] }
passlib = { version = "^1.7", extras = ["argon2"] }
# Async support
aiosqlite = "^0.19"
asyncpg = { version = "^0.28", optional = true }
# APScheduler (standalone, not Flask version)
APScheduler = "^3.10"
# Common dependencies
future = "^0.18"
itsdangerous = "^2.1"
markdown = "^3.3"
multiqc = "^1.9"
numpy = ">=1.14,<2.0"
plotly = ">2,<6"
pyyaml = ">=5.1"
# Pydantic for validation (comes with FastAPI)
pydantic = { version = "^2.0", extras = ["email"] }
pydantic-settings = "^2.0"
# Marshmallow for serialization
marshmallow = "^3.0"
marshmallow-sqlalchemy = ">=0.29"
marshmallow-jsonapi = "^0.24"
outlier-utils = "^0.0.3"
querystring-parser = "^1.2"
scipy = "^1.5"
flatten_json = "^0.1"
environs = "^9.2.0"
scikit-learn = "^1.2.0"
marshmallow-jsonschema = "^0.13.0"
marshmallow-polyfield = "^5.11"
httpx = { version = ">=0.24", optional = true }
pytest = { version = "^7.2", optional = true }
pytest-asyncio = { version = "^0.21", optional = true }
factory-boy = { version = ">2.12,<4.0", optional = true }
livereload = { version = "^2.5", optional = true }
pre-commit = { version = "*", optional = true }
wheel = { version = "^0.30", optional = true }
psycopg2 = { version = "^2.6", optional = true }

[tool.poetry.extras]
dev = [
        "pytest",
        "pytest-asyncio",
        "httpx",
        "factory-boy",
        "livereload",
        "pre-commit",
]
deploy = ["wheel"]
prod = ["psycopg2", "asyncpg"]

[tool.poetry.scripts]
megaqc = "megaqc.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
